Create web implementation of ShopAPIPlugin
