+++ web/src/ShopAPIPlugin/index.ts
@@ -1,0 +1,2 @@
+export { default as ShopAPI } from './ShopAPIPlugin';
+export * from './definitions';
